EKK-LM3S9B96: Rename files for better conformance to naming conventions
