+++ nuttx/configs/ekk-lm3s9b96/src/lm_boot.c
@@ -1,8 +1,7 @@
 /************************************************************************************
- * configs/lm3s6965-ek/src/up_ethernet.c
- * arch/arm/src/board/up_ethernet.c
+ * configs/ekk-lm3s9b96/src/lm_boot.c
  *
- *   Copyright (C) 2012 Gregory Nutt. All rights reserved.
+ *   Copyright (C) 2012, 2015 Gregory Nutt. All rights reserved.
  *   Authors: Gregory Nutt <gnutt@nuttx.org>
  *            Jose Pablo Rojas V. <jrojas@nx-engineering.com>
  *
@@ -41,16 +40,14 @@
 
 #include <nuttx/config.h>
 
-#include <stdint.h>
 #include <debug.h>
-#include <assert.h>
 
+#include <nuttx/board.h>
 #include <arch/board/board.h>
-#include <net/ethernet.h>
 
 #include "up_arch.h"
-#include "chip.h"
-#include "tiva_ethernet.h"
+#include "up_internal.h"
+#include "ekklm3s9b96_internal.h"
 
 /************************************************************************************
  * Pre-processor Definitions
@@ -65,36 +62,31 @@
  ************************************************************************************/
 
 /************************************************************************************
- * Name: tiva_ethernetmac
+ * Name: tiva_boardinitialize
  *
  * Description:
- *   For the Ethernet Eval Kits, the MAC address will be stored in the non-volatile
- *   USER0 and USER1 registers.  If CONFIG_TIVA_BOARDMAC is defined, this function
- *   will obtain the MAC address from these registers.
+ *   All Stellaris architectures must provide the following entry point.  This entry
+ *   point is called early in the intitialization -- after all memory has been
+ *   configured and mapped but before any devices have been initialized.
  *
  ************************************************************************************/
 
-#ifdef CONFIG_TIVA_BOARDMAC
-void tiva_ethernetmac(struct ether_addr *ethaddr)
+void tiva_boardinitialize(void)
 {
-  uint32_t user0;
-  uint32_t user1;
+  /* Configure chip selects if 1) SSI is not disabled, and 2) the weak function
+   * lm_ssiinitialize() has been brought into the link.
+   */
 
-  /* Get the current value of the user registers */
-
-  user0 = getreg32(TIVA_FLASH_USERREG0);
-  user1 = getreg32(TIVA_FLASH_USERREG1);
-
-  nlldbg("user: %06x:%06x\n", user1 & 0x00ffffff, user0 & 0x00ffffff);
-  DEBUGASSERT(user0 != 0xffffffff && user1 != 0xffffffff);
-
-  /* Re-format that MAC address the way that uIP expects to see it */
-
-  ethaddr->ether_addr_octet[0] = ((user0 >>  0) & 0xff);
-  ethaddr->ether_addr_octet[1] = ((user0 >>  8) & 0xff);
-  ethaddr->ether_addr_octet[2] = ((user0 >> 16) & 0xff);
-  ethaddr->ether_addr_octet[3] = ((user1 >>  0) & 0xff);
-  ethaddr->ether_addr_octet[4] = ((user1 >>  8) & 0xff);
-  ethaddr->ether_addr_octet[5] = ((user1 >> 16) & 0xff);
-}
+#if defined(CONFIG_TIVA_SSI0)  || defined(CONFIG_TIVA_SSI1)
+  if (lm_ssiinitialize)
+    {
+      lm_ssiinitialize();
+    }
 #endif
+
+  /* Configure on-board LEDs if LED support has been selected. */
+
+#ifdef CONFIG_ARCH_LEDS
+  board_led_initialize();
+#endif
+}
