+++ nuttx/configs/ekk-lm3s9b96/src/lm_ethernet.c
@@ -1,7 +1,7 @@
 /************************************************************************************
- * configs/ekk-lm3s9b96/src/up_boot.c
+ * configs/lm3s6965-ek/src/lm_ethernet.c
  *
- *   Copyright (C) 2012, 2015 Gregory Nutt. All rights reserved.
+ *   Copyright (C) 2012 Gregory Nutt. All rights reserved.
  *   Authors: Gregory Nutt <gnutt@nuttx.org>
  *            Jose Pablo Rojas V. <jrojas@nx-engineering.com>
  *
@@ -40,14 +40,16 @@
 
 #include <nuttx/config.h>
 
+#include <stdint.h>
 #include <debug.h>
+#include <assert.h>
 
-#include <nuttx/board.h>
 #include <arch/board/board.h>
+#include <net/ethernet.h>
 
 #include "up_arch.h"
-#include "up_internal.h"
-#include "ekklm3s9b96_internal.h"
+#include "chip.h"
+#include "tiva_ethernet.h"
 
 /************************************************************************************
  * Pre-processor Definitions
@@ -62,31 +64,36 @@
  ************************************************************************************/
 
 /************************************************************************************
- * Name: tiva_boardinitialize
+ * Name: tiva_ethernetmac
  *
  * Description:
- *   All Stellaris architectures must provide the following entry point.  This entry
- *   point is called early in the intitialization -- after all memory has been
- *   configured and mapped but before any devices have been initialized.
+ *   For the Ethernet Eval Kits, the MAC address will be stored in the non-volatile
+ *   USER0 and USER1 registers.  If CONFIG_TIVA_BOARDMAC is defined, this function
+ *   will obtain the MAC address from these registers.
  *
  ************************************************************************************/
 
-void tiva_boardinitialize(void)
+#ifdef CONFIG_TIVA_BOARDMAC
+void tiva_ethernetmac(struct ether_addr *ethaddr)
 {
-  /* Configure chip selects if 1) SSI is not disabled, and 2) the weak function
-   * lm_ssiinitialize() has been brought into the link.
-   */
+  uint32_t user0;
+  uint32_t user1;
 
-#if defined(CONFIG_TIVA_SSI0)  || defined(CONFIG_TIVA_SSI1)
-  if (lm_ssiinitialize)
-    {
-      lm_ssiinitialize();
-    }
-#endif
+  /* Get the current value of the user registers */
 
-  /* Configure on-board LEDs if LED support has been selected. */
+  user0 = getreg32(TIVA_FLASH_USERREG0);
+  user1 = getreg32(TIVA_FLASH_USERREG1);
 
-#ifdef CONFIG_ARCH_LEDS
-  board_led_initialize();
-#endif
+  nlldbg("user: %06x:%06x\n", user1 & 0x00ffffff, user0 & 0x00ffffff);
+  DEBUGASSERT(user0 != 0xffffffff && user1 != 0xffffffff);
+
+  /* Re-format that MAC address the way that uIP expects to see it */
+
+  ethaddr->ether_addr_octet[0] = ((user0 >>  0) & 0xff);
+  ethaddr->ether_addr_octet[1] = ((user0 >>  8) & 0xff);
+  ethaddr->ether_addr_octet[2] = ((user0 >> 16) & 0xff);
+  ethaddr->ether_addr_octet[3] = ((user1 >>  0) & 0xff);
+  ethaddr->ether_addr_octet[4] = ((user1 >>  8) & 0xff);
+  ethaddr->ether_addr_octet[5] = ((user1 >> 16) & 0xff);
 }
+#endif
