+++ nuttx/configs/ekk-lm3s9b96/src/lm_ssi.c
@@ -1,6 +1,5 @@
 /************************************************************************************
- * configs/ekk-lm3s9b96/src/up_ssi.c
- * arch/arm/src/board/up_ssi.c
+ * configs/ekk-lm3s9b96/src/lm_ssi.c
  *
  *   Copyright (C) 2012 Gregory Nutt. All rights reserved.
  *   Authors: Gregory Nutt <gnutt@nuttx.org>
@@ -56,7 +55,7 @@
 #if defined(CONFIG_TIVA_SSI0)  || defined(CONFIG_TIVA_SSI1)
 
 /************************************************************************************
- * Definitions
+ * Pre-processor Definitions
  ************************************************************************************/
 
 /* Enables debug output from this file (needs CONFIG_DEBUG too) */
